remove py2cpp: gather_nd (#6889)
